Add fix for ODK Intent return values
